test(power-supply): split test file under 500-LOC file-size budget
CI's file-size-budget job rejected the original 541-line test_power_supply.py
(budget is 500 LOC per changed file). Split into two test files at a natural
boundary:

- test_power_supply.py (249 LOC) — config validation, controller
  construction, permissive transitions, mode switching, status snapshot,
  math invariants, update_config behavior
- test_power_supply_runtime.py (297 LOC) — feedback-driven behavior:
  current/power setpoint clamping + validation, HH_POWER/HH_TEMP trip
  latching, acknowledge, tick() output guarantees

Extracted shared factory helpers (fresh_idle / fresh_armed / fresh_running)
into a single underscore-prefixed module so pytest doesn't try to collect it
as a test file. All 57 tests still pass.
